# Speedup XML parsing