--- a/lib/pmdtester/parsers/pmd_report_document.rb
+++ b/lib/pmdtester/parsers/pmd_report_document.rb
@@ -58,8 +58,10 @@ module PmdTester
         @violations.add_violations_by_filename(@current_filename, @current_violations)
         @current_filename = nil
       when 'violation'
-        @current_violation.text.strip!
-        @current_violations.push(@current_violation) if match_filter_set?(@current_violation)
+        if match_filter_set?(@current_violation)
+          @current_violation.text.strip!
+          @current_violations.push(@current_violation)
+        end
         @current_violation = nil
       when 'error'
         @errors.add_error_by_filename(@current_filename, @current_error)
@@ -79,15 +81,12 @@ module PmdTester
     def match_filter_set?(violation)
       return true if @filter_set.nil?
 
-      @filter_set.each do |filter_rule_ref|
-        ruleset_attr = violation.attrs['ruleset'].delete(' ').downcase + '.xml'
-        rule = violation.rule_name
-        rule_ref = "#{ruleset_attr}/#{rule}"
-        return true if filter_rule_ref.eql?(ruleset_attr)
-        return true if filter_rule_ref.eql?(rule_ref)
-      end
+      ruleset_attr = violation.attrs['ruleset'].delete(' ').downcase + '.xml'
+      return true if @filter_set.include?(ruleset_attr)
 
-      false
+      rule_ref = "#{ruleset_attr}/#{violation.rule_name}"
+
+      @filter_set.include?(rule_ref)
     end
   end
 end